refactor(toolkit-lib): extract pure helpers out of resource-investigation
resource-investigation.ts had grown to mix orchestration (the AWS-calling
investigate/fetch/resolve functions) with a pile of pure utilities. Move the
pure helpers to dedicated homes, leaving resource-investigation focused on
orchestration (908 -> 681 lines):

- CloudWatch log formatting (trimToRecentLines, parseLambdaLogEvents + helpers,
  cloudWatchLogsConsoleUrl, MAX_LOG_LINES) -> format-utils.ts
- Resource identifier parsers (parseEcsServiceIdentifier,
  serviceTokenReferencedLogicalId, functionNameFromArnOrName, extractLogStreamName,
  logStreamNameFromPhysicalId, ecsStoppedTasksConsoleUrl) -> new resource-identifiers.ts

Behavior-preserving; tests only change import paths.
